+++ packages/block-editor/src/components/block-list/use-in-between-inserter.js
@@ -76,8 +76,6 @@ export function useInBetweenInserter() {
 					rootClientId = blockElement.getAttribute( 'data-block' );
 				}
 
-				// Don't show the in-between inserter if the container block is
-				// locked or disabled.
 				if (
 					getTemplateLock( rootClientId ) ||
 					getBlockEditingMode( rootClientId ) === 'disabled'